Refactor job handling: move job-related functions to apps.go, clean up apps_run.go and cluster.go
